@@ -1,17 +1,17 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 32
     defaultConfig {
         applicationId "mhashim6.android.thetwoauthentics"
-        minSdkVersion 15
-        targetSdkVersion 26
-        versionCode 30
-        versionName "1.9.9.5"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        minSdkVersion 16
+        targetSdkVersion 32
+        versionCode 31
+        versionName "1.9.9.6"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         vectorDrawables.useSupportLibrary = true
         resConfigs "en", "ar"
-        buildToolsVersion "28.0.3"
+        buildToolsVersion "31.0.0"
     }
     buildTypes {
         release {
@@ -30,20 +30,21 @@ android {
 }
 
 dependencies {
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    implementation  'com.android.support:appcompat-v7:28.0.0'
-    implementation  'com.android.support:recyclerview-v7:28.0.0'
-    implementation  'com.android.support:cardview-v7:28.0.0'
-    implementation  'com.android.support:design:28.0.0'
-    implementation  'com.android.support:support-v4:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.5.1'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'com.google.android.material:material:1.6.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation  'com.readystatesoftware.sqliteasset:sqliteassethelper:2.0.1'
-    implementation  'uk.co.chrisjenx:calligraphy:2.2.0'
+    implementation 'io.github.inflationx:calligraphy3:3.1.1'
+    implementation 'io.github.inflationx:viewpump:2.0.3'
     implementation  'com.github.javiersantos:AppUpdater:2.6.1'
-    implementation  'com.android.support:customtabs:28.0.0'
+    implementation 'androidx.browser:browser:1.4.0'
     implementation  'com.github.hotchemi:android-rate:1.0.1'
-    testImplementation 'junit:junit:4.12'
+    testImplementation 'junit:junit:4.13.2'
 }
 repositories {
     mavenCentral()
