@@ -2,7 +2,7 @@ package mhashim6.android.thetwoauthentics.app;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import static mhashim6.android.thetwoauthentics.app.Utils.WORKERS;
 
