@@ -1,6 +1,6 @@
 package mhashim6.android.thetwoauthentics.app.results;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import java.util.Collections;
 import java.util.List;
